+++ compiler/src/group/edwards_bls12.rs
@@ -16,7 +16,7 @@ use snarkos_models::{
             boolean::Boolean,
             eq::{ConditionalEqGadget, EqGadget},
             select::CondSelectGadget,
-            uint8::UInt8,
+            uint::UInt8,
             ToBitsGadget, ToBytesGadget,
         },
     },
